frame update
played around with the jframe and cleared out errors!!
